fix fix  :log up bug ,create translate swither
