+++ my-app/src/components/Header/index.tsx
@@ -16,13 +16,19 @@ import { Link } from 'react-router-dom';
 import { useAppSelector, RootState, useAppDispatch } from 'store/store';
 import { Button } from '@mui/material';
 import { logout } from 'store/auth/auth-slice';
+import { useTranslation, Trans } from 'react-i18next';
 
 const settings = [
   { name: 'Sign In', link: 'sign-in' },
   { name: 'Sign Up', link: 'sign-up' },
 ];
+const lngs = {
+  en: { nativeName: 'English' },
+  ru: { nativeName: 'Russia' },
+};
 
 function Header() {
+  const { t, i18n } = useTranslation();
   const userId: string = useAppSelector((state: RootState) => state.rootReducer.authReducer.userId);
   const dispatch = useAppDispatch();
   const [anchorElUser, setAnchorElUser] = React.useState<null | HTMLElement>(null);
@@ -95,14 +101,18 @@ function Header() {
           >
             {userId ? (
               <Link to="/boards">
-                <Typography>BOARDS</Typography>
+                <Typography>
+                  <Trans i18nKey="header.boards"></Trans>
+                </Typography>
               </Link>
             ) : (
               <></>
             )}
             {userId ? (
               <Link to="/UserPage">
-                <Typography>EDIT PROFILE</Typography>
+                <Typography>
+                  <Trans i18nKey="header.edit"></Trans>
+                </Typography>
               </Link>
             ) : (
               <></>
@@ -155,8 +165,17 @@ function Header() {
               marginLeft: '15px',
             }}
           >
-            <ToggleButton value="EN">EN</ToggleButton>
-            <ToggleButton value="RU">RU</ToggleButton>
+            {Object.keys(lngs).map((lng) => (
+              <ToggleButton
+                value={lng}
+                type="submit"
+                key={lng}
+                onClick={() => i18n.changeLanguage(lng)}
+                disabled={i18n.resolvedLanguage === lng}
+              >
+                {lngs[lng as keyof typeof lngs].nativeName}
+              </ToggleButton>
+            ))}
           </ToggleButtonGroup>
         </Toolbar>
       </Container>
